Refactor auth guards
