+++ frontend/src/app/app-routing.module.ts
@@ -3,7 +3,7 @@ import { Routes, RouterModule } from '@angular/router';
 
 import { HomePageComponent } from './home-page/home-page.component';
 import { LoggedOutGuard } from './modules/auth/guards/logged-out.guard';
-import { LoggedInGuard } from './modules/auth/guards/logged-in-guard.service';
+import { LoggedInGuard } from './modules/auth/guards/logged-in.guard';
 import { PublicFormComponent } from './modules/public-form/public-form/public-form.component';
 import { LoginComponent } from './modules/auth/login/login.component';
