+++ frontend/src/app/modules/auth/guards/logged-in.guard.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed, async, inject } from '@angular/core/testing';
 
-import { LoggedInGuard } from './logged-in-guard.service';
+import { LoggedInGuard } from './logged-in.guard';
 
 describe('LoggedInGuard', () => {
   beforeEach(() => {
