+++ frontend/src/app/modules/auth/guards/logged-out.guard.ts
@@ -1,27 +1,21 @@
-import { Injectable } from '@angular/core';
+import {Injectable} from '@angular/core';
 import {CanActivate, Router} from '@angular/router';
-import { Observable } from 'rxjs';
-import {map, take} from 'rxjs/operators';
-import { AuthModule } from '../auth.module';
-import { AuthService } from '../auth.service';
+import {Observable} from 'rxjs';
+import {map, take, tap} from 'rxjs/operators';
+import {AuthModule} from '../auth.module';
+import {AuthService} from '../auth.service';
 
 @Injectable({
     providedIn: AuthModule,
 })
 export class LoggedOutGuard implements CanActivate {
-    constructor(private authService: AuthService, private router: Router) { }
+    constructor(private authService: AuthService, private router: Router) {}
 
     canActivate(): Observable<boolean> {
         return this.authService.isLoggedIn$.pipe(
             take(1),
-            map(isUserLoggedIn => {
-                if (isUserLoggedIn) {
-                    this.router.navigate(['interwencje/zglos']);
-                    return false;
-                } else {
-                    return true;
-                }
-            })
+            tap(isLoggedIn => isLoggedIn ? this.router.navigate(['interwencje/zglos']) : null),
+            map(isLoggedIn => !isLoggedIn),
         );
     }
 }
